Do some reorganization and fix an error handling check
Moved most of the sql stuff to the common functions file.

Instead of checking that the count is null I'm now checking that the
count is neither null or undefined.
