+++ routes.js
@@ -1,32 +1,7 @@
 var express = require('express');
 var common = require('./lib/commonFunctions');
-var Sequelize = require('sequelize');
-var sequelize = new Sequelize('engagement', 'engagement', 'password', {
-        dialect: 'sqlite',
-        pool: {
-            max: 5,
-            min: 0,
-            idle: 10000,
-        },
-        storage: __dirname + "/database.sqlite"
-});
 
-var Song = sequelize.define('song', {
-    title: {
-        type: Sequelize.STRING,
-        unique: 'composite'
-    },
-    artist: {
-        type: Sequelize.STRING,
-        unique: 'composite'
-    },
-    count: {
-        type: Sequelize.INTEGER
-    }
-});
-
-Song.sync();
-
+common.setupDatabase();
 
 var router = express.Router();
 router.get(/^\/(story)?$/, common.renderTemplate('story'));
@@ -37,7 +12,7 @@ router.get('/groomsmen', common.renderTemplate('groomsmen'));
 router.get('/weddinginfo', common.renderTemplate('weddinginfo'));
 router.get('/gallery', common.renderTemplate('gallery'));
 router.get('/songsuggestions', function (req, res) {
-    Song.findAll().then(function(songs) {
+    common.Song.findAll().then(function(songs) {
         res.json(songs);
     });
 });
@@ -50,23 +25,9 @@ router.post('/songsuggestions', function (req, res) {
         });
         return;
     }
-    Song.sync().then(function() {
-        Song.findOrCreate({
-            where: {
-                title: title,
-                artist: artist
-            }
-        }).then(function(song) {
-            var actualSong = song[0];
-            if(actualSong.count === null) {
-                actualSong.count = 0;
-            }
-            actualSong.count++;
-            actualSong.save();
-            res.json(actualSong);
-        });
+    common.Song.addSong(title, artist).then(function(song) {
+        res.json(song);
     });
-
 });
 
 module.exports = router;
